feat(hypervisor): Ontology Manager + Object Explorer module extraction + shared ontology context — zero drift (#59)
The Ontology Application Runtime wave opens with its runway: both certified
ontology ports leave the serve monolith as app modules, and the shared
semantic-layer context kit lands (deliberately unwired — PR60-62 wire it).

Modules (render code moved VERBATIM, zero behavior change by construction):
- surfaces/ontology-manager/index.mjs — slug schema, /__ioi/ontology/manager
- surfaces/object-explorer/index.mjs — slug explorer, /__ioi/ontology/explorer
Both under the surface contract {meta, load(ctx), render(model, ctx), actions:[]},
sharing one loader (loadOntologyModel — the exact four ODK projections the flat
branches fetched). The registry imports and binds both; the serve's two route
branches and both inline renderers are deleted (~42k chars). Routes unchanged.

ZERO DRIFT proven: manager and explorer HTML byte-identical before/after,
including the ?ontology= and ?q= variants (4/4 diffs empty). Cert artifacts
untouched; pixel harness 37/37.

Shared ontology context (surfaces/ontology-context.mjs): parseOntologyContext/
ontologyContextQuery over the six semantic keys (ontology/objectType/objectSet/
objectId/pane/section; canonical order, roundtrip-stable; reuses the kit's
parse/serialize — no duplicated logic, no second escaper), managerLink/
explorerLink/objectTypeLink/objectSetLink, semanticBreadcrumb
(Ontology → Object type → Object set → Source/Pipeline/Proof),
semanticInspectorShell, disabledSemanticAction, formatRef.

surface-modules verifier extended (+20 checks, 43/43): both modules' contract/
registry/matrix agreement + identity binding + offline dead-daemon renders
keeping certified landmarks + serve hygiene (no ontology port renderers remain;
the odk substrate's own renderer stays) + full context-kit unit coverage.

Also: two stale asserts in the ontology-manager verifier repinned to the
CURRENT certified rail markup (og-nlabel label wrap from the pixel-cert
restyle) — proven pre-existing against the pre-extraction baseline bytes; the
semantic checks are unchanged (count pill cross-check, named gaps disabled).

Done bar: manager 23/23, explorer 25/25, surface-modules 43/43, pipeline 77/77,
pixel 37/37, runtime-safety 20/20, reset 23/23, approvals 23/23, models 17/17,
matrix clean, certs untouched, no NUL, no Rust.
