+++ apps/hypervisor/scripts/serve-product-ui.mjs
@@ -5324,425 +5324,11 @@ function renderOntologyManager(ov, lists, selectedId) {
   return automationsShell("Ontology Manager", body);
 }
 
-// ============================ ONTOLOGY MANAGER — reference UX PORT (#34, daemon_wired).
-// A FAITHFUL source-neutral port of the reference Ontology Manager UX (NOT a dark native redesign):
-//   - a DARK global platform RAIL (source-neutral IOI nav: Home/Search/…/Applications · Ontology Manager)
-//   - a LIGHT app RAIL: Discover / Proposals / History · Resources (Object types · Properties · Link
-//     types · Action types · Value types · Functions) · Health issues / Cleanup / Ontology configuration
-//   - a LIGHT HEADER: app title · ontology switcher · "Search resources…" · New
-//   - a LIGHT card-first BODY: "Object types recently modified" as object-type CARDS, then the typed
-//     schema detail + configuration below.
-// Light theme + card-first IA + the reference's landmark labels — matched so the HARDENED harness
-// (theme + IA landmarks + region geometry, #34 review) can certify visual parity. Wired to the REAL
-// ODK CanonicalObjectModel; READ-ONLY (authoring + object materialization stay in /__ioi/odk).
-
 // ---- SHARED GLOBAL RAIL: MOVED VERBATIM to surfaces/chrome.mjs (imported above) — the rail is
 // aligned once and certified in every port shell; it now lives with the surface modules.
-
-function renderOntologyManagerPort(ov, lists, selectedId) {
-  const enc = encodeURIComponent, esc = CX_ESC;
-  const o = ov || {};
-  const ontologies = Array.isArray(lists.ontologies) ? lists.ontologies : [];
-  const selected = ontologies.find((x) => x.id === selectedId) || ontologies.find((x) => (x.health || {}).status === "ready") || ontologies[0] || null;
-  const com = (selected && selected.canonical_object_model) || {};
-  const arr = (k) => (Array.isArray(com[k]) ? com[k] : []);
-  const vts = arr("value_types"), ots = arr("object_types"), lts = arr("link_types"), ats = arr("action_types");
-  const funcs = ats.filter((a) => a.kind === "function"), nonFuncActs = ats.filter((a) => a.kind !== "function");
-  const health = (selected && selected.health) || {};
-  const msets = (lists.materialized_sets || []).filter((m) => selected && m.ontology_ref === selected.ref);
-  const projs = (lists.projections || []).filter((p) => selected && p.ontology_ref === selected.ref);
-  const totalInstances = msets.reduce((a, m) => a + (m.count || 0), 0);
-  const propCount = ots.reduce((n, x) => n + (Array.isArray(x.properties) ? x.properties.length : 0), 0);
-  const rollup = o.ontology_health || {};
-  const idc = (x) => `<code class="om-code">${esc(x || "")}</code>`;
-  const hstate = health.status || "empty";
-  const hpill = `<span class="om-pill ${hstate === "ready" ? "ok" : hstate === "empty" ? "muted" : "warn"}">${esc(hstate)}</span>`;
-
-  const oid = selected ? selected.id : "";
-  const domainLabel = selected ? esc(selected.domain || selected.id) : "no ontology";
-  const objsOf = (t) => msets.filter((m) => m.object_type_id === t.id).reduce((a, m) => a + (m.count || 0), 0);
-  const depsOf = (t) => lts.filter((l) => l.from === t.id || l.to === t.id).length;
-  const svg = (p) => `<svg viewBox="0 0 24 24" width="17" height="17" fill="none" stroke="currentColor" stroke-width="1.7" stroke-linecap="round" stroke-linejoin="round">${p}</svg>`;
-  const CUBE = '<path d="M12 2l9 5v10l-9 5-9-5V7z"/>';
-
-  // DARK global platform rail — the SHARED pixel-aligned reference shell (ioiGlobalRailHtml).
-  const globalRail = ioiGlobalRailHtml({ label: "Ontology Manager", href: "/__ioi/ontology/manager", iconUri: ONTOLOGY_APP_ICON_URI });
-
-  // LIGHT header — the reference navbar layout: app-icon chip · title · centered Search-resources input
-  // (ctrl+K) · branch selector (named gap) · New. The ONTOLOGY SWITCHER (a live control with no reference
-  // counterpart in the navbar) lives in the BODY section heading — the shell stays pixel-faithful while
-  // the live control keeps its function where the ontology name already appears.
-  const header = `<header class="og-header">
-    <span class="og-hchip"></span>
-    <h5 class="og-htitle">Ontology Manager</h5>
-    <div class="og-search" title="Search + ctrl+K command palette are reference-only — not wired (named gap)"><span class="og-sico">${bpIcon("search")}</span><input placeholder="Search resources…" disabled aria-label="Search resources (reference-only, not wired)"><span class="og-skbd">ctrl + K</span></div>
-    <div class="og-headright">
-      <span class="og-branch" title="Branching is a reference-only lane — no authority contract yet (named gap).">${bpIcon("git-branch")}<span class="og-branchname">Main</span>${bpIcon("caret-down")}</span>
-      <a class="og-new" href="/__ioi/odk/ontologies/new">New ${bpIcon("caret-down")}</a>
-    </div>
-  </header>`;
-
-  // LIGHT app rail — the reference nav column: icon rows at a 40px pitch, sentence-case "Resources"
-  // header, indented resource items with right-aligned COUNT PILLS (live daemon values — masked as
-  // dynamic data on the IOI side; the reference's captured counts are masked on its side), hairline
-  // group dividers, Health/Cleanup/Configuration at group level.
-  const arailItem = (icon, label, count, href, opts = {}) => {
-    const c = count == null ? "" : `<span class="og-c">${count}</span>`;
-    const inner = `<span class="og-nico">${bpIcon(icon)}</span><span class="og-nlabel">${esc(label)}</span>${c}`;
-    const cls = `og-nav${opts.on ? " on" : ""}${opts.sub ? " sub" : ""}`;
-    return href
-      ? `<a class="${cls}" href="${href}">${inner}</a>`
-      : `<span class="${cls} gap" title="${esc(label)} is a reference-only lane — no authority contract yet (named gap).">${inner}</span>`;
-  };
-  const appRail = `<nav class="og-arail" aria-label="Ontology Manager">
-    ${arailItem("compass", "Discover", null, "#og-discover", { on: true })}
-    ${arailItem("people", "Proposals", null, null)}
-    ${arailItem("time", "History", null, "#og-config")}
-    <div class="og-adiv"></div>
-    <div class="og-asec">Resources</div>
-    ${arailItem("cube", "Object types", ots.length, "#og-discover", { sub: true })}
-    ${arailItem("properties", "Properties", propCount, "#og-properties", { sub: true })}
-    ${arailItem("globe-network", "Shared properties", 0, null, { sub: true })}
-    ${arailItem("arrows-horizontal", "Link types", lts.length, "#og-link-types", { sub: true })}
-    ${arailItem("take-action", "Action types", nonFuncActs.length, "#og-action-types", { sub: true })}
-    ${arailItem("group-objects", "Groups", 0, null, { sub: true })}
-    ${arailItem("selection-box", "Interfaces", 0, null, { sub: true })}
-    <div class="og-agap"></div>
-    ${arailItem("intersection", "Value types", vts.length, "#og-value-types", { sub: true })}
-    ${arailItem("function", "Functions", funcs.length, "#og-functions", { sub: true })}
-    <div class="og-adiv2"></div>
-    ${arailItem("pulse", "Health issues", null, "#og-health")}
-    ${arailItem("clean", "Cleanup", null, null)}
-    ${arailItem("cog", "Ontology configuration", null, "#og-config")}
-  </nav>`;
-
-  // LIGHT card-first body: object-type cards ("recently modified"), then typed detail + configuration.
-  const cardOf = (t) => {
-    const desc = (t.description || "").trim();
-    return `<a class="og-card" href="/__ioi/odk/ontologies/${selected ? enc(selected.id) : "new"}#ot-${enc(t.id)}" title="${esc(t.name || t.id)} — open in the substrate ontology detail">
-      <div class="og-cardtop"><span class="og-cardico">${svg(CUBE)}</span><span class="og-cardname">${esc(t.name || t.id)}</span><span class="og-cardbook">${svg('<path d="M4 5a2 2 0 012-2h12v18H6a2 2 0 01-2-2z"/><path d="M8 3v18"/>')}</span></div>
-      <div class="og-cardobj"><b>${objsOf(t)}</b> object${objsOf(t) === 1 ? "" : "s"}</div>
-      <div class="og-carddep">${depsOf(t)} dependent${depsOf(t) === 1 ? "" : "s"}</div>
-      <div class="og-carddesc">${desc ? esc(desc) : "No description"}</div>
-    </a>`;
-  };
-  const tbl = (head, rows, empty) => rows ? `<table class="og-table"><thead><tr>${head.map((h) => `<th>${h}</th>`).join("")}</tr></thead><tbody>${rows}</tbody></table>` : `<div class="og-none">${empty}</div>`;
-  const propRows = ots.flatMap((t) => (Array.isArray(t.properties) ? t.properties : []).map((p) => `<tr><td>${esc(t.name || t.id)}</td><td>${esc(p.name || p.id)} ${idc(p.id)}</td><td>${idc(p.value_type || "")}</td><td>${p.required ? "yes" : "—"}${t.title_property === p.id ? ` <span class="om-pill ok">title</span>` : ""}</td></tr>`)).join("");
-  const valRows = vts.map((v) => `<tr><td>${esc(v.name || v.id)} ${idc(v.id)}</td><td><span class="om-pill muted">${esc(v.base || "string")}</span></td><td>${(v.enum_values && v.enum_values.length) ? v.enum_values.map((e) => `<span class="om-pill muted">${esc(e)}</span>`).join(" ") : "—"}</td></tr>`).join("");
-  const linkRows = lts.map((l) => `<tr><td>${esc(l.name || l.id)} ${idc(l.id)}</td><td>${idc(l.from || "")} → ${idc(l.to || "")}</td><td><span class="om-pill muted">${esc(l.cardinality || "")}</span></td></tr>`).join("");
-  const actRows = nonFuncActs.map((a) => `<tr><td>${esc(a.name || a.id)} ${idc(a.id)}</td><td><span class="om-pill muted">${esc(a.kind || "")}</span></td><td>${a.applies_to ? idc(a.applies_to) : "—"}</td></tr>`).join("");
-  const funcRows = funcs.map((a) => `<tr><td>${esc(a.name || a.id)} ${idc(a.id)}</td><td>${a.applies_to ? idc(a.applies_to) : "—"}</td></tr>`).join("");
-  const body = `<main class="og-body" role="main">${selected ? `
-    <section id="og-discover" class="og-discover">
-      <div class="og-sechd"><h2>Object types recently modified in</h2>
-        <details class="og-ontomenu"><summary>${domainLabel}${selected ? ` <span class="og-ver">v${esc(selected.version || "0.1.0")}</span>` : ""} ▾</summary>
-          <div class="og-ontolist">${ontologies.length ? ontologies.map((x) => `<a class="og-ontoitem${selected && x.id === selected.id ? " on" : ""}" href="/__ioi/ontology/manager?ontology=${enc(x.id)}">${esc(x.domain || x.id)} <span class="og-dot og-${(x.health || {}).status === "ready" ? "ok" : (x.health || {}).status === "empty" ? "muted" : "warn"}"></span></a>`).join("") : `<div class="og-none">No ontologies yet.</div>`}</div>
-        </details>
-        <a class="og-explorerlink" href="/__ioi/ontology/explorer" title="Object Explorer — browse object types and materialized object sets (the symmetric #35 surface)">Object Explorer →</a>
-        <a class="og-configlink" href="/__ioi/odk/ontologies/${enc(selected.id)}/edit">Configure</a></div>
-      ${ots.length ? `<div class="og-cards">${ots.map(cardOf).join("")}</div>` : `<div class="og-none">No object types yet. <a href="/__ioi/odk/ontologies/${enc(selected.id)}/edit">Add typed object types →</a></div>`}
-    </section>
-    <section id="og-properties"><h2>Properties <span class="og-subn">${propCount}</span></h2>${tbl(["Object type", "Property", "Value type", "Required"], propRows, "No properties declared.")}</section>
-    <section id="og-value-types"><h2>Value types <span class="og-subn">${vts.length}</span></h2>${tbl(["Value type", "Base", "Enum"], valRows, "No value types.")}</section>
-    <section id="og-link-types"><h2>Link types <span class="og-subn">${lts.length}</span></h2>${tbl(["Link", "From → To", "Cardinality"], linkRows, "No link types.")}</section>
-    <section id="og-action-types"><h2>Action types <span class="og-subn">${nonFuncActs.length}</span></h2><div class="og-note">Action <b>declarations</b> only — writeback/execution is not wired (needs a PolicyBoundDataView + TransformationRun). Declaring an action never runs it.</div>${tbl(["Action", "Kind", "Applies to"], actRows, "No action types.")}</section>
-    <section id="og-functions"><h2>Functions <span class="og-subn">${funcs.length}</span></h2><div class="og-note">Function <b>declarations</b> only — evaluation/execution is not wired.</div>${tbl(["Function", "Applies to"], funcRows, "No function declarations.")}</section>
-    <section id="og-health"><h2>Health issues</h2><div class="og-healthbox"><div class="og-sechd2"><b>Readiness</b> ${hpill} <span class="og-sub">${(health.counts || {}).object_types || 0} obj · ${(health.counts || {}).value_types || 0} val · ${(health.counts || {}).link_types || 0} link · ${(health.counts || {}).action_types || 0} act</span></div>${(health.gaps || []).length ? `<ul class="og-gaps">${health.gaps.map((g) => `<li>${esc(g)}</li>`).join("")}</ul>` : `<div class="og-sub">No health issues — the required semantic pieces are present.</div>`}<div class="og-sub" style="margin-top:6px"><b>${esc(String(health.object_instances == null ? 0 : health.object_instances))}</b> object instances — ${esc(health.object_data_note || "schema only; no object-instance plane bound until an OntologyProjection exists.")}</div></div></section>
-    <section id="og-config"><h2>Ontology configuration</h2><div class="og-cfg">
-      <div class="og-cfgrow"><span>Ref</span>${idc(selected.ref)}</div>
-      <div class="og-cfgrow"><span>Revision</span><b>rev ${esc(String(selected.revision || 1))}</b></div>
-      <div class="og-cfgrow"><span>Receipts</span><b>${(selected.receipt_refs || []).length}</b></div>
-      <div class="og-cfgrow"><span>History</span><b>${(selected.history || []).length}</b></div>
-      <div class="og-cfgrow"><span>Object sets · objects</span><b>${msets.length} · ${totalInstances}</b></div>
-      <div class="og-cfgrow"><span>Projections</span><b>${projs.length}</b></div>
-      <div class="og-cfgrow"><span>Estate</span><span><span class="om-pill ok">${rollup.ready || 0} ready</span> <span class="om-pill warn">${rollup.incomplete || 0} incomplete</span> <span class="om-pill muted">${rollup.empty || 0} empty</span></span></div>
-      <a class="og-editlink" href="/__ioi/odk/ontologies/${enc(selected.id)}/edit">Configure model in substrate →</a>
-      <div class="og-note" style="margin-top:8px">Named gaps (reference-only lanes, no authority contract yet): in-canvas schema editing · Proposals · Shared properties · Groups · Interfaces · Cleanup · action/function execution. Reference: <a href="/__apps/schema" target="_blank" rel="noopener">Ontology Manager ↗</a>.</div>
-    </div></section>
-  ` : `<div class="og-none" style="margin:40px auto;max-width:520px">Select or create an ontology to see its schema. <a href="/__ioi/odk/ontologies/new">Create an ontology →</a></div>`}</main>`;
-
-  const css = `@font-face{font-family:'Source-Sans-Pro';font-style:normal;font-weight:400;font-display:block;src:url(/__ioi/fonts/source-sans-pro-400.woff2) format('woff2')}
-    @font-face{font-family:'Source-Sans-Pro';font-style:normal;font-weight:600;font-display:block;src:url(/__ioi/fonts/source-sans-pro-600.woff2) format('woff2')}
-    @font-face{font-family:'Source-Sans-Pro';font-style:normal;font-weight:700;font-display:block;src:url(/__ioi/fonts/source-sans-pro-700.woff2) format('woff2')}
-    html{color-scheme:light}*{box-sizing:border-box}
-    body{margin:0;background:#f4f5f7;color:#1c2127;font:14px/1.28581 Source-Sans-Pro,Helvetica,sans-serif}
-    a{color:#2f6fd8;text-decoration:none}
-    .og-shell{display:flex;height:100vh;width:100vw;overflow:hidden}
-    ${IOI_GRAIL_CSS}
-    .og-main{flex:1;min-width:0;display:flex;flex-direction:column;height:100vh}
-    .og-header{flex:0 0 auto;position:relative;height:50px;display:flex;align-items:center;padding:0 15px 0 0;background:#fff;box-shadow:0 1px 0 0 #dce0e5;z-index:5}
-    .og-hchip{width:50px;height:50px;flex:0 0 50px;background:rgba(102,158,255,.1) url('${ONTOLOGY_APP_ICON_URI}') center no-repeat}
-    .og-htitle{margin:0 0 0 15px;font-weight:600;font-size:16px;color:#1c2127;line-height:18px}
-    .og-ontomenu{position:relative;display:inline-block}.og-ontomenu>summary{list-style:none;cursor:pointer;font-size:14px;color:#1c2127;background:#f1f3f6;border:1px solid #e0e3e8;border-radius:4px;padding:3px 10px}
-    .og-ontomenu>summary::-webkit-details-marker{display:none}.og-ver{color:#8b9099}
-    .og-ontolist{position:absolute;top:32px;left:0;min-width:230px;background:#fff;border:1px solid #e0e3e8;border-radius:9px;box-shadow:0 8px 28px rgba(20,24,31,.14);padding:6px;z-index:20}
-    .og-ontoitem{display:flex;align-items:center;justify-content:space-between;padding:6px 9px;border-radius:6px;color:#3a3f46}
-    .og-ontoitem:hover{background:#f1f3f6}.og-ontoitem.on{background:#eef2fb;color:#1a1d21;font-weight:600}
-    .og-search{position:absolute;left:50%;top:10px;transform:translateX(-50%);width:350px;height:30px;display:flex;align-items:center;background:#f6f7f9;border-radius:4px;padding:0 10px 0 7px}
-    .og-sico{display:inline-flex;color:#5f6b7c;flex:0 0 16px;margin-right:1px}
-    .og-search input{flex:1;border:0;background:transparent;font:inherit;font-size:14px;line-height:16px;height:16px;padding:0;color:#1c2127;outline:none}
-    .og-search input::placeholder{color:#5f6b7c}
-    .og-skbd{font-size:14px;line-height:16px;color:#5f6b7c;white-space:nowrap}
-    .og-headright{margin-left:auto;display:flex;align-items:center;gap:23px}
-    .og-branch{display:inline-flex;align-items:center;gap:8px;color:#5f6b7c;font-size:14px;line-height:16px;cursor:default}
-    .og-branchname{color:#1c2127}
-    .og-new{display:inline-flex;align-items:center;justify-content:center;gap:8px;width:70px;height:30px;padding:0;border:1px solid #d7dade;border-radius:4px;background:#fff;color:#1c2127;font-size:14px;line-height:16px;font-weight:400}
-    .og-new svg{color:#5f6b7c}
-    .og-work{flex:1 1 auto;display:flex;min-height:0}
-    .og-arail{flex:0 0 299px;width:299px;background:#fff;border-right:1px solid #dce0e5;overflow-y:auto;padding:6px 5px 6px 6px}
-    .og-nav{display:flex;align-items:center;gap:10px;height:35px;padding:0 10px;border-radius:3px;color:#1c2127;font-size:14px;margin:0 0 5px}
-    .og-nav.sub{padding-left:15px}
-    .og-nico{display:inline-flex;align-items:center;justify-content:center;width:16px;height:16px;color:#5f6b7c;flex:0 0 16px}
-    .og-nlabel{flex:1 1 auto;white-space:nowrap;overflow:hidden;text-overflow:ellipsis}
-    .og-nav:hover{background:#f6f7f9}
-    .og-nav.on{background:#f3f8ff;color:#215db0}.og-nav.on .og-nico{color:#215db0}
-    .og-nav .og-c{font-size:12px;color:#1c2127;background:#eef0f3;border-radius:4px;padding:1px 7px;line-height:18px}
-    .og-nav.gap{color:#1c2127;cursor:default}.og-nav.gap:hover{background:transparent}
-    .og-asec{font-size:14px;color:#1c2127;padding:1px 10px 9px 15px;font-weight:600}
-    .og-adiv{height:1px;background:#e5e8eb;margin:10px 10px}
-    .og-adiv2{height:1px;background:#e5e8eb;margin:5px 10px 10px}
-    .og-agap{height:14px}
-    .og-body{flex:1 1 auto;overflow:auto;padding:20px 24px;background:#f6f7f9}
-    .og-body section{margin:0 0 26px}
-    .og-sechd{display:flex;align-items:center;justify-content:space-between;margin:0 0 12px}
-    .og-body h2{font-size:15px;margin:0 0 12px;font-weight:600}.og-subn{color:#9aa0a8;font-weight:400;font-size:13px}
-    .og-configlink{font-size:13px}
-    .og-explorerlink{font-size:13px;margin-right:14px}
-    .og-cards{display:grid;grid-template-columns:repeat(auto-fill,minmax(280px,1fr));gap:14px}
-    .og-card{display:block;background:#fff;border:1px solid #e4e7ec;border-radius:10px;padding:14px 15px;box-shadow:0 1px 2px rgba(20,24,31,.04);color:#1a1d21}
-    .og-card:hover{border-color:#c9d3e6;box-shadow:0 3px 10px rgba(20,24,31,.08)}
-    .og-cardtop{display:flex;align-items:center;gap:9px;margin:0 0 8px}.og-cardico{display:inline-flex;color:#5b6472}
-    .og-cardname{font-weight:600;font-size:13.5px;flex:1;overflow:hidden;text-overflow:ellipsis;white-space:nowrap}.og-cardbook{color:#b6bcc4}
-    .og-cardobj{font-size:12.5px;color:#3a3f46}.og-carddep{font-size:12px;color:#8b9099;margin:2px 0 8px}
-    .og-carddesc{font-size:12.5px;color:#9aa0a8;font-style:italic}
-    .og-table{border-collapse:collapse;width:100%;font-size:12.5px;background:#fff;border:1px solid #e6e8ec;border-radius:8px;overflow:hidden}
-    .og-table th{text-align:left;color:#8b9099;font-weight:600;padding:8px 14px;border-bottom:1px solid #eceef1;background:#fafbfc}
-    .og-table td{padding:8px 14px;border-bottom:1px solid #f0f1f4}
-    .og-note{color:#6b7178;font-size:12px;border:1px solid #e4e7ec;border-radius:8px;padding:9px 12px;margin:0 0 10px;background:#fbfcfd}
-    .og-healthbox,.og-cfg{background:#fff;border:1px solid #e6e8ec;border-radius:10px;padding:14px 16px}
-    .og-sechd2{display:flex;align-items:center;gap:8px;margin:0 0 6px}.og-gaps{margin:6px 0 0;padding-left:18px;color:#6b7178}
-    .og-cfgrow{display:flex;justify-content:space-between;gap:10px;padding:6px 0;border-bottom:1px solid #f0f1f4}.og-cfgrow>span:first-child{color:#8b9099}
-    .og-editlink{display:inline-block;margin-top:10px;font-size:13px}
-    .og-none{color:#8b9099;padding:16px;border:1px dashed #d8dbe0;border-radius:10px;background:#fff}
-    .og-sub{color:#8b9099;font-size:12px}
-    .om-code{font-family:ui-monospace,monospace;font-size:11px;color:#6b7178;background:#f1f3f6;padding:1px 5px;border-radius:4px}
-    .og-dot{width:8px;height:8px;border-radius:50%;display:inline-block}
-    .og-dot.og-ok{background:#22a35a}.og-dot.og-warn{background:#d6a13a}.og-dot.og-muted{background:#aab0b8}
-    .om-pill{display:inline-block;padding:1px 8px;border-radius:999px;font-size:11px;border:1px solid;white-space:nowrap}
-    .om-pill.ok{color:#1a7f43;border-color:#bfe4cd;background:#eafaf0}
-    .om-pill.warn{color:#a2730c;border-color:#efd9a6;background:#fdf6e6}
-    .om-pill.muted{color:#6b7178;border-color:#e0e3e8;background:#f3f4f6}`;
-
-  return `<!doctype html><html><head><meta charset="utf-8"><meta name="viewport" content="width=device-width,initial-scale=1"><title>Ontology Manager</title><style>${css}</style></head>
-    <body><div class="og-shell">${globalRail}<div class="og-main">${header}<div class="og-work">${appRail}${body}</div></div></body></html>`;
-}
-
-// ============================ OBJECT EXPLORER — reference UX PORT (#35, reference_ported).
-// A FAITHFUL source-neutral port of the reference Object Explorer (dark global platform rail + a light
-// "Object Explorer search" header with the Filter/Search bar + a Shortcuts strip + an Object type
-// CATALOG table + an Object set CATALOG), wired to the REAL ODK truth (object types across ontologies,
-// materialized object sets, per-type object + usage counts, a working server-side object-type filter).
-// READ-ONLY; the sibling /__ioi/ontology/manager is linked first-class. NOT daemon_wired: the local
-// /workspace/hubble reference does not cleanly boot (the proxy renders a blank body; the mirror's data
-// lanes render "Failed to load"), so the hardened harness has no valid reference to certify
-// visual_parity against — honest `reference_ported`, promotable on a clean re-harvest.
-// ============================ OBJECT EXPLORER (#46 — the origin-aligned promotion. #44 proved the
-// old "blank/failed Hubble reference" blocker WRONG: the capture-origin lane localhost:9225/workspace/
-// hubble/ renders the full Object Explorer with data. This port is the faithful shell of THAT
-// reference over real IOI ODK truth: object types across live DomainOntologies, object counts from
-// materialized sets, the object-set catalog from real materialized sets, and a WORKING ?q= type
-// filter. Reference-only lanes (object-instance search, Filter-by facets, Recents/Favorites, sort,
-// type-group/application lanes, exploration tabs, ontology selector) are named gaps disabled in
-// place. The catalog/set ROWS are the live body (excluded from shell-pixel certification, verified
-// semantically); the chrome is glyph-anchored to the reference: tab bar h40 · centered search hero ·
-// shortcuts row + cards · catalog heading/filter/sort band · table header · object-set band. The
-// content block is the reference's responsive rule: max-width 1400, width calc(100% − 120px),
-// centered (margins 60 @1440 → 145 @1920).
-function renderObjectExplorerPort(ov, lists, opts) {
-  const enc = encodeURIComponent, esc = CX_ESC;
-  const q = (opts && opts.q ? String(opts.q) : "").trim();
-  const ontologies = Array.isArray(lists.ontologies) ? lists.ontologies : [];
-  const msets = Array.isArray(lists.materialized_sets) ? lists.materialized_sets : [];
-  const projs = Array.isArray(lists.projections) ? lists.projections : [];
-  const arr = (oo, k) => { const com = (oo && oo.canonical_object_model) || {}; return Array.isArray(com[k]) ? com[k] : []; };
-
-  // REAL daemon truth: the flat object-type catalog + per-type materialized counts + link usage.
-  const allTypes = ontologies.flatMap((oo) => arr(oo, "object_types").map((t) => ({ oo, t })));
-  const objectsOf = (oo, t) => msets.filter((m) => m.ontology_ref === oo.ref && m.object_type_id === t.id).reduce((a, m) => a + (m.count || 0), 0);
-  const usageOf = (oo, t) => arr(oo, "link_types").filter((l) => l.from === t.id || l.to === t.id).length;
-  const catalog = q ? allTypes.filter(({ oo, t }) => `${t.name || ""} ${t.id || ""} ${oo.domain || ""}`.toLowerCase().includes(q.toLowerCase())) : allTypes;
-  const fmtN = (n) => (n >= 1000 ? `${Math.round(n / 100) / 10}K` : String(n));
-  // Shortcuts = the real top materialized sets by object count (live data — masked in the pixel gate).
-  const shortcuts = msets.slice().sort((a, b) => (b.count || 0) - (a.count || 0)).slice(0, 3);
-  const CHIP_COLORS = ["#bd6bbd", "#13c9ba", "#4c90f0"];
-
-  const typeRow = ({ oo, t }) => {
-    const n = objectsOf(oo, t);
-    return `<tr class="oe-trow" onclick="location.href='/__ioi/ontology/manager?ontology=${enc(oo.id)}'">
-      <td class="oe-tname"><span class="oe-tchip" style="color:${CHIP_COLORS[(t.id || "").length % 3]}">${bpIcon("cube", 14)}</span>${esc(t.name || t.id)}</td>
-      <td class="oe-tstatus">${bpIcon("manual", 14)}</td>
-      <td>${esc(fmtN(n))}</td>
-      <td>${usageOf(oo, t)} link${usageOf(oo, t) === 1 ? "" : "s"}</td>
-      <td class="oe-tgroups"></td>
-      <td class="oe-tdesc">${esc(oo.domain || oo.id)}</td>
-    </tr>`;
-  };
-  const setRow = (m) => `<tr class="oe-trow" onclick="location.href='/__ioi/odk?ontology=${enc((ontologies.find((oo) => oo.ref === m.ontology_ref) || {}).id || "")}'">
-    <td class="oe-tname"><span class="oe-tchip" style="color:#13c9ba">${bpIcon("layout-grid", 14)}</span>${esc(m.name || m.set_id || m.object_type_id)}</td>
-    <td>${esc(fmtN(m.count || 0))} object${(m.count || 0) === 1 ? "" : "s"}</td>
-    <td class="oe-tdesc">${esc((ontologies.find((oo) => oo.ref === m.ontology_ref) || {}).domain || m.ontology_ref || "")}</td>
-    <td class="oe-tdesc"><a href="/__ioi/ontology/manager">Ontology Manager →</a></td>
-  </tr>`;
-
-  const globalRail = ioiGlobalRailHtml({ label: "Object Explorer", href: "/__ioi/ontology/explorer", iconUri: EXPLORER_APP_ICON_URI, railVariant: "rv-pipe", viewAll: false, star: false, badges: true, aipGradient: true, acctMuted: true, hiliteNav: "Ontology" });
-
-  const tabbar = `<div class="oe-tabbar oe-topbar">
-    <span class="oe-sqbtn gap" aria-disabled="true" title="Active exploration — a reference-only session lane (named gap)"><span class="oe-sqico"></span></span>
-    <span class="oe-tab" title="Explorations are a reference-only session lane — this tab is the honest default state">${bpIcon("search")}<span class="oe-tabt">New exploration</span></span>
-    <span class="oe-plus gap" aria-disabled="true" title="Opening more exploration tabs is a reference-only lane (named gap)">${bpIcon("plus")}</span>
-    <span class="oe-ontsel gap" aria-disabled="true" title="Ontology scoping is a reference-only lane — the catalog below spans ALL live ontologies honestly">All Ontologies ${bpIcon("caret-down")}</span>
-  </div>`;
-
-  const hero = `<div class="oe-hero">
-    <h2 class="oe-htitle">Object Explorer search</h2>
-    <div class="oe-searchrow">
-      <div class="oe-herogrp">
-        <span class="oe-filterby gap" aria-disabled="true" title="Faceted object filters are a reference-only lane — no object-instance search plane (named gap)">${bpIcon("filter-funnel")}<span class="oe-fbt">Filter by...</span>${bpIcon("caret-down")}</span>
-        <div class="oe-objsearch" title="Full-text object search is a reference-only lane — objects exist as materialized sets (the catalog below is real)">${bpIcon("search")}<input placeholder="Search for objects..." disabled aria-label="Search for objects (reference-only, not wired)"><span class="oe-send gap" aria-disabled="true">${bpIcon("send-to")}</span></div>
-      </div>
-    </div>
-  </div>`;
-
-  const cards = shortcuts.map((m, i) => `<a class="oe-card" href="/__ioi/odk?ontology=${enc((ontologies.find((oo) => oo.ref === m.ontology_ref) || {}).id || "")}">
-    <span class="oe-cchip" style="background:${CHIP_COLORS[i]}1a;color:${CHIP_COLORS[i]}">${bpIcon("layout-grid", 14)}</span>
-    <span class="oe-cbody"><span class="oe-ctitle">${esc(m.name || m.set_id || m.object_type_id)}</span><span class="oe-csub">Object Type&nbsp;&nbsp;•&nbsp;&nbsp;${esc(fmtN(m.count || 0))} object${(m.count || 0) === 1 ? "" : "s"}</span></span>
-  </a>`).join("");
-  const shortcutsBand = `<div class="oe-shrow">
-    <span class="oe-shlabel">Shortcuts</span>
-    <span class="oe-lanes">
-      <span class="oe-lane on gap" aria-disabled="true" title="Recents are a reference-only per-user lane (named gap) — the cards below are the real top materialized sets">Recents</span>
-      <span class="oe-lane gap" aria-disabled="true" title="Favorites are a reference-only per-user lane (named gap)">Favorites</span>
-      <span class="oe-lane gap" aria-disabled="true" title="Your object sets — the real materialized sets render in the catalog below">Your object sets</span>
-      <span class="oe-shchev gap" aria-disabled="true">${bpIcon("chevron-right")}</span>
-    </span>
-  </div>
-  <div class="oe-cards">${cards || `<div class="oe-cardempty">No materialized sets yet — run a materializing run to populate shortcuts. This row reads real daemon sets; nothing is fabricated.</div>`}</div>`;
-
-  const catalogBand = `<h3 class="oe-cathead"><span class="oe-cathd">Object type catalog</span></h3>
-  <div class="oe-filterrow">
-    <form class="oe-filterform" method="GET" action="/__ioi/ontology/explorer">${bpIcon("filter-funnel")}<input name="q" value="${esc(q)}" placeholder="Filter for an object type..." aria-label="Filter for an object type (live)"><span class="oe-count">${catalog.length} of ${allTypes.length}</span></form>
-    <span class="oe-sortlanes">
-      <span class="oe-sort gap" aria-disabled="true" title="Relevancy sorting is a reference-only lane — rows are ordered by ontology then name (named gap)">${bpIcon("sort-desc")}<span class="oe-sortt">Relevancy</span>${bpIcon("caret-down")}</span>
-      <span class="oe-lane on gap" aria-disabled="true" title="named gap">All</span>
-      <span class="oe-lane gap" aria-disabled="true" title="Type groups are a reference-only lane — the daemon records none (named gap)">Type group</span>
-      <span class="oe-lane gap" aria-disabled="true" title="Application scoping is a reference-only lane (named gap)">Application</span>
-    </span>
-  </div>
-  <div class="oe-tablebox">
-    <table class="oe-table">
-      <thead><tr><th class="oe-thname">Object type name</th><th class="oe-thstatus">Status</th><th class="oe-thcount">Object count</th><th class="oe-thusage">Usage</th><th class="oe-thgroups">Type groups</th><th>Description</th></tr></thead>
-      <tbody>${catalog.length ? catalog.map(typeRow).join("") : `<tr><td colspan="6" class="oe-none">${q ? `No object types match “${esc(q)}”.` : "No object types yet — create an ontology in the Ontology Manager."} <a href="/__ioi/ontology/manager">Ontology Manager →</a></td></tr>`}</tbody>
-    </table>
-  </div>`;
-
-  const setBand = `<div class="oe-setrow">
-    <span class="oe-setlabel">Object set catalog <span class="oe-setsub">(explorations and lists)</span></span>
-    <span class="oe-setlanes">
-      <input class="oe-setsearch" placeholder="Search explorations..." disabled aria-label="Search explorations (reference-only, not wired)" title="Exploration search is a reference-only lane (named gap)">
-      <span class="oe-slane on gap" aria-disabled="true" title="named gap">All</span>
-      <span class="oe-slane gap" aria-disabled="true" title="Per-user set lanes are reference-only (named gap)">Created by me</span>
-      <span class="oe-slane gap" aria-disabled="true" title="named gap">Shared with me</span>
-      <span class="oe-slane gap" aria-disabled="true" title="named gap">Favorites</span>
-    </span>
-  </div>
-  <div class="oe-setbox">
-    <table class="oe-table oe-settable">
-      <tbody>${msets.length ? msets.slice(0, 20).map(setRow).join("") : `<tr><td class="oe-none">No object sets yet — a materialized set appears once an OntologyProjection reads a source (${projs.length} projection${projs.length === 1 ? "" : "s"} declared). <a href="/__ioi/odk">ODK substrate →</a></td></tr>`}</tbody>
-    </table>
-    <div class="oe-foot">Every row is daemon truth: ${allTypes.length} object type${allTypes.length === 1 ? "" : "s"} across ${ontologies.length} live ontolog${ontologies.length === 1 ? "y" : "ies"} · ${msets.length} materialized set${msets.length === 1 ? "" : "s"}. Schema authoring: <a href="/__ioi/ontology/manager">Ontology Manager →</a> · substrate: <a href="/__ioi/odk">ODK</a> · reference: <a href="/__apps/explorer" target="_blank" rel="noopener">Object Explorer capture ↗</a></div>
-  </div>`;
-
-  const css = `:root{color-scheme:light}*{box-sizing:border-box}
-    body{margin:0;background:#fff;color:#1c2127;font:14px/1.28581 Source-Sans-Pro,Helvetica,sans-serif}
-    a{color:#215db0;text-decoration:none}
-    .oe-shell{display:flex;height:100vh;width:100vw;overflow:hidden}
-    ${IOI_GRAIL_CSS}
-    .oe-main{flex:1;min-width:0;display:flex;flex-direction:column;height:100vh}
-    .oe-tabbar{flex:0 0 40px;display:flex;align-items:stretch;background:#edeff2;box-shadow:inset 0 -1px 0 0 rgba(17,20,24,.15)}
-    .oe-sqbtn{width:41px;display:flex;align-items:center;justify-content:center;background:rgba(189,173,255,.1)}
-    .oe-sqico{width:24px;height:24px;border-radius:4px;background:rgba(167,182,194,.1) url('${EXPLORER_APP_ICON_URI}') center/20px no-repeat}
-    .oe-tab{display:flex;align-items:center;gap:5px;width:180px;padding:0 0 0 15px;background:#fff;box-shadow:inset 0 -1px 0 0 rgba(17,20,24,.15)}
-    .oe-tab svg{color:#5f6b7c}
-    .oe-tabt{font-size:14px;line-height:18.0013px;color:#1c2127}
-    .oe-plus{display:flex;align-items:center;justify-content:center;width:30px;color:#5f6b7c}
-    .oe-ontsel{margin-left:auto;display:inline-flex;align-items:center;gap:8px;height:30px;margin-top:5px;margin-right:10px;padding:0 8px;border-radius:4px;font-size:14px;line-height:16.1px;color:#1c2127;cursor:default;background:#fff;box-shadow:inset 0 0 0 1px rgba(64,72,84,.33),0 1px 2px rgba(17,20,24,.1)}
-    .oe-ontsel svg{color:#5f6b7c}
-    .oe-body{flex:1 1 auto;min-width:0;overflow-y:auto;overflow-x:hidden;background:#fff}
-    .oe-content{max-width:1400px;width:calc(100% - 121px);margin:0 auto;position:relative}
-    .oe-hero{text-align:center}
-    .oe-htitle{font-size:16px;line-height:19px;font-weight:600;color:#1c2127;margin:21px 0 0;position:relative;top:-1px}
-    .oe-searchrow{display:flex;justify-content:center;margin-top:16px}
-    .oe-herogrp{display:flex;border-radius:4px;box-shadow:0 0 0 1px rgba(17,20,24,.17),0 2px 8px rgba(17,20,24,.22);margin-left:2px}
-    .oe-filterby{display:inline-flex;align-items:center;width:250px;height:40px;padding:0 12px;background:#f6f7f9;border-right:1px solid rgba(17,20,24,.15);border-radius:4px 0 0 4px;cursor:default}
-    .oe-filterby svg{color:#5f6b7c}
-    .oe-fbt{flex:1;font-size:14px;line-height:18.0013px;color:#5f6b7c;margin-left:4px;text-align:left}
-    .oe-objsearch{display:flex;align-items:center;width:450px;height:40px;padding:0 5px 0 8px;background:#fff;border-radius:0 4px 4px 0}
-    .oe-objsearch svg{color:#5f6b7c}
-    .oe-objsearch input{flex:1;border:0;background:transparent;font:inherit;font-size:14px;line-height:18.0013px;color:#1c2127;outline:none;padding:0;margin-left:4px}
-    .oe-objsearch input::placeholder{color:#5f6b7c}
-    .oe-send{display:inline-flex;align-items:center;justify-content:center;width:32px;height:32px;border-radius:50%;background:#2d72d2}
-    .oe-send svg{color:#fff}
-    .oe-shrow{display:flex;align-items:center;margin-top:19px}
-    .oe-shlabel{font-size:14px;line-height:16px;font-weight:600;color:#1c2127}
-    .oe-lanes{margin-left:auto;display:inline-flex;align-items:center}
-    .oe-lane{display:inline-flex;align-items:center;height:30px;padding:0 8px;border-radius:4px;font-size:14px;line-height:16.1px;color:#5f6b7c;cursor:default;margin-left:2px}
-    .oe-lane.on{background:#fff;color:#1c2127;box-shadow:inset 0 0 0 1px rgba(64,72,84,.33),0 1px 2px rgba(17,20,24,.1)}
-    .oe-shchev{display:inline-flex;color:#5f6b7c;margin:0 7.5px 0 26px}
-    .oe-cards{display:flex;gap:14px;margin-top:11px}
-    .oe-card{flex:1;display:flex;align-items:center;height:57px;padding:0 14px;border-radius:4px;box-shadow:inset 0 0 0 1px rgba(17,20,24,.15)}
-    .oe-cchip{display:inline-flex;align-items:center;justify-content:center;width:24px;height:24px;border-radius:4px;flex:0 0 24px}
-    .oe-cbody{display:flex;flex-direction:column;margin-left:14px;min-width:0}
-    .oe-ctitle{font-size:14px;line-height:18.0013px;color:#1c2127;white-space:nowrap;overflow:hidden;text-overflow:ellipsis}
-    .oe-csub{font-size:14px;line-height:18.0013px;color:#5f6b7c}
-    .oe-cardempty{flex:1;display:flex;align-items:center;height:57px;padding:0 14px;border-radius:4px;border:1px dashed #d3d8de;color:#5f6b7c;font-size:12.5px}
-    .oe-cathead{margin:63px 0 0;font-size:14px;line-height:16px;font-weight:600;color:#1c2127}
-    .oe-cathd{border-bottom:1px dotted #8f99a8;line-height:15px;display:inline-block}
-    .oe-filterrow{display:flex;align-items:center;margin-top:12px}
-    .oe-filterform{display:flex;align-items:center;width:450px;height:30px;padding:0 5px 0 7px;background:#fff;box-shadow:inset 0 0 0 1px rgba(17,20,24,.2),inset 0 1px 1px rgba(17,20,24,.3);border-radius:4px}
-    .oe-filterform svg{color:#5f6b7c}
-    .oe-filterform input{flex:1;border:0;background:transparent;font:inherit;font-size:14px;color:#1c2127;outline:none;padding:0;margin-left:7px}
-    .oe-filterform input::placeholder{color:#5f6b7c}
-    .oe-count{display:inline-flex;align-items:center;height:20px;padding:0 6px;border-radius:4px;background:rgba(143,153,168,.15);font-size:12px;line-height:16px;color:#1c2127;white-space:nowrap}
-    .oe-sortlanes{margin-left:auto;display:inline-flex;align-items:center}
-    .oe-sort{display:inline-flex;align-items:center;gap:8px;height:30px;padding:0 8px;font-size:14px;line-height:16.1px;color:#1c2127;cursor:default;margin-right:12px}
-    .oe-sort svg{color:#5f6b7c}
-    .oe-table{border-collapse:collapse;width:100%;font-size:14px;table-layout:fixed}
-    .oe-tablebox{margin-top:8px;height:430px;overflow-y:auto;border-radius:4px;box-shadow:0 0 0 1px rgba(17,20,24,.15)}
-    .oe-table thead th{position:sticky;top:0;background:#f6f7f9;text-align:left;font-size:12px;line-height:15.4297px;font-weight:400;color:#5f6b7c;text-transform:uppercase;padding:7.3px 10px 8px;border-bottom:1px solid #e5e8eb;white-space:nowrap;overflow:hidden;text-overflow:ellipsis}
-    .oe-thname{width:284.8px;padding-left:20px !important}
-    .oe-thstatus{width:79.2px}.oe-thcount{width:90.8px}.oe-thusage{width:181.5px}.oe-thgroups{width:181.5px}
-    .oe-trow{cursor:pointer}
-    .oe-trow:hover{background:#f6f7f9}
-    .oe-trow td{padding:11px 10px;border-bottom:1px solid #eef0f2;color:#1c2127;line-height:18.0013px;white-space:nowrap;overflow:hidden;text-overflow:ellipsis;max-width:280px}
-    .oe-tname{padding-left:20px !important}
-    .oe-tchip{display:inline-flex;vertical-align:-2px;margin-right:12px}
-    .oe-tstatus svg{color:#935610}
-    .oe-tdesc{color:#5f6b7c !important}
-    .oe-none{padding:18px 20px;color:#5f6b7c}
-    .oe-setrow{display:flex;align-items:center;margin-top:24px}
-    .oe-setlabel{font-size:14px;line-height:16px;font-weight:600;color:#1c2127}
-    .oe-setsub{font-weight:400;color:#5f6b7c}
-    .oe-setlanes{margin-left:auto;display:inline-flex;align-items:center;gap:8px}
-    .oe-setsearch{width:219px;height:30px;border:0;border-radius:30px;background:#fff;box-shadow:inset 0 0 0 1px rgba(17,20,24,.2);font:inherit;font-size:14px;color:#1c2127;outline:none;padding:0 12px 0 30px}
-    .oe-slane{display:inline-flex;align-items:center;height:30px;padding:0 10px;border-radius:30px;font-size:14px;line-height:18.0013px;color:#1c2127;cursor:default}
-    .oe-slane.on{background:#c0d4f1;color:#184a90;font-weight:600}
-    .oe-setbox{margin:10px 0 30px}
-    .oe-foot{margin-top:14px;color:#7b8494;font-size:12px;line-height:1.5}`;
-
-  return `<!doctype html><html><head><meta charset="utf-8"><meta name="viewport" content="width=device-width,initial-scale=1"><title>Object Explorer</title><style>${css}</style></head>
-    <body><div class="oe-shell">${globalRail}<div class="oe-main">${tabbar}<div class="oe-body"><main class="oe-content" role="main">${hero}${shortcutsBand}${catalogBand}${setBand}</main></div></div></div></body></html>`;
-}
-
+// ============================ ONTOLOGY MANAGER + OBJECT EXPLORER: EXTRACTED (Ontology wave) ======
+// Both certified ports live as surface modules now — surfaces/ontology-manager/index.mjs and
+// surfaces/object-explorer/index.mjs — mounted through the surface registry like Pipeline.
 function domainAppPayload(p) {
   const csv = (k) => (p.get(k) || "").split(",").map((s) => s.trim()).filter(Boolean);
   const payload = {
@@ -9706,34 +9292,6 @@ async function handleEstateRequest(req, res, body) {
     }
     // Ontology Manager — reference UX PORT (#34, daemon_wired). Ported schema-workbench shell over the
     // real ODK CanonicalObjectModel; the /__ioi/odk substrate/authoring surface stays as-is.
-    if (pathname === "/__ioi/ontology/manager" && req.method === "GET") {
-      const J = (p) => fetch(`${DAEMON}${p}`).then((r) => r.json()).catch(() => ({}));
-      const [ov, o, op, ms] = await Promise.all([
-        J("/v1/hypervisor/odk/overview"),
-        J("/v1/hypervisor/odk/domain-ontologies"),
-        J("/v1/hypervisor/odk/ontology-projections"),
-        J("/v1/hypervisor/odk/materialized-object-sets"),
-      ]);
-      const selectedOntology = new URL(req.url, "http://x").searchParams.get("ontology") || "";
-      res.writeHead(200, HTMLH);
-      res.end(renderOntologyManagerPort(ov, { ontologies: o.ontologies || [], projections: op.ontology_projections || [], materialized_sets: ms.materialized_object_sets || [] }, selectedOntology));
-      return;
-    }
-    // Object Explorer — reference UX PORT (#35, reference_ported). Faithful Object Explorer shell over
-    // the real ODK object-type + materialized-set truth; NOT daemon_wired (blank/failed Hubble reference).
-    if (pathname === "/__ioi/ontology/explorer" && req.method === "GET") {
-      const J = (p) => fetch(`${DAEMON}${p}`).then((r) => r.json()).catch(() => ({}));
-      const [ov, o, op, ms] = await Promise.all([
-        J("/v1/hypervisor/odk/overview"),
-        J("/v1/hypervisor/odk/domain-ontologies"),
-        J("/v1/hypervisor/odk/ontology-projections"),
-        J("/v1/hypervisor/odk/materialized-object-sets"),
-      ]);
-      const q = new URL(req.url, "http://x").searchParams.get("q") || "";
-      res.writeHead(200, HTMLH);
-      res.end(renderObjectExplorerPort(ov, { ontologies: o.ontologies || [], projections: op.ontology_projections || [], materialized_sets: ms.materialized_object_sets || [] }, { q }));
-      return;
-    }
     if (pathname === "/__ioi/odk" && req.method === "GET") {
       const J = (p) => fetch(`${DAEMON}${p}`).then((r) => r.json()).catch(() => ({}));
       const [ov, o, r, d, m, ds, cm, pv, tr, op, lp, mr, cs, ms] = await Promise.all([
